+++ extensions/polymarket/CHANGELOG.md
@@ -4,3 +4,6 @@
 Removed 2024 US Election Odds and Swing States
 
 ## [Initial Version] - 2024-11-05
+
+## [PolyMarket Search] 2025 -05-10
+- Added ability to search PolyMarket for a specific market.